+++ src/filter/index.ts
@@ -1,2 +1,6 @@
+export type {
+    FilterItemCheckboxOptionProps,
+    FilterItemCheckboxProps,
+} from "./addons/types";
 export * from "./filter";
 export * from "./types";
