+++ stories/filter/filter.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as FilterStories from "./filter.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={FilterStories} />
 
@@ -83,4 +83,4 @@ Some filters may need to rendered as a fullscreen page while editing. A display 
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={FilterStories} />
